refactor: move monitor_runners.go to usecase root
Move monitor_runners.go from usecase/monitor/ to usecase/ directory
and remove unnecessary monitor subdirectory. Update package name
from 'monitor' to 'usecase' and fix all import references.
