+++ internal/presentation/tui/model.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/VeyronSakai/gh-runner-monitor/internal/domain/entity"
 	"github.com/VeyronSakai/gh-runner-monitor/internal/domain/value_object"
-	"github.com/VeyronSakai/gh-runner-monitor/internal/usecase/monitor"
+	"github.com/VeyronSakai/gh-runner-monitor/internal/usecase"
 	"github.com/charmbracelet/bubbles/table"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
@@ -14,7 +14,7 @@ import (
 // Model represents the TUI application state
 type Model struct {
 	table          table.Model
-	useCase        *monitor.MonitorRunnersUseCase
+	useCase        *usecase.MonitorRunnersUseCase
 	owner          string
 	repo           string
 	org            string
@@ -29,7 +29,7 @@ type Model struct {
 }
 
 // NewModel creates a new TUI model
-func NewModel(useCase *monitor.MonitorRunnersUseCase, owner, repo, org string) *Model {
+func NewModel(useCase *usecase.MonitorRunnersUseCase, owner, repo, org string) *Model {
 	columns := []table.Column{
 		{Title: "Runner Name", Width: 25},
 		{Title: "Status", Width: 12},
